chore(refactor): rename AppActions -> AppAction cause why not
